API: make hutch a required positional argument for the GUI.
In the py2 version, this was a --hutch mandatory argument. Prior to this commit it was a --hutch optional argument (secretly mandatory, would raise without it). However, the iocmanager wrapper script converts this to a positional hutch argument. This change synchronizes us with the wrapper script so that when we show the help text we don't confuse the user.